broadwell: Create iomap.h header with platform base addresses
Instead of having various defined base addresses in different
headers put them all in one well defined location.  The names
are changed to be more consistent with baytrail implementation.

Some defines are for early temporary base addresses, a few of
which are taken from Kconfig variables and are set here in order
to be consistent with the ones that are not defined in Kconfig.

The code will be changed to use the new defines in subsequent
commits.

BUG=chrome-os-partner:28234
TEST=None

Change-Id: I315d8c6f4188244bc86342e8c5dce60924653c58
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/198736
Reviewed-by: Aaron Durbin <adurbin@chromium.org>